允许实时调用mp4decrpyt解密

+++ src/N_m3u8DL-RE/DownloadManager/SimpleDownloadManager.cs
@@ -51,6 +51,15 @@ namespace N_m3u8DL_RE.DownloadManager
             var headers = DownloaderConfig.Headers;
             var output = Path.Combine(saveDir, saveName + $".{streamSpec.Extension ?? "ts"}");
 
+            //mp4decrypt
+            var APP_DIR = Path.GetDirectoryName(Environment.ProcessPath)!;
+            var fileName = "mp4decrypt";
+            if (Environment.OSVersion.Platform == PlatformID.Win32NT)
+                fileName += ".exe";
+            var mp4decrypt = Path.Combine(APP_DIR, fileName);
+            if (!File.Exists(mp4decrypt)) mp4decrypt = fileName;
+            var mp4InitFile = "";
+
             Logger.Debug($"dirName: {dirName}; tmpDir: {tmpDir}; saveDir: {saveDir}; saveName: {saveName}; output: {output}");
 
             //创建文件夹
@@ -76,12 +85,20 @@ namespace N_m3u8DL_RE.DownloadManager
                 var path = Path.Combine(tmpDir, "_init.mp4.tmp");
                 var result = await Downloader.DownloadSegmentAsync(streamSpec.Playlist.MediaInit, path, headers);
                 FileDic[streamSpec.Playlist.MediaInit] = result;
+                if (result == null)
+                {
+                    throw new Exception("Download init file failed!");
+                }
+                mp4InitFile = result.ActualFilePath;
                 task.Increment(1);
+
                 //修改输出后缀
                 if (streamSpec.MediaType == Common.Enum.MediaType.AUDIO)
                     output = Path.ChangeExtension(output, ".m4a");
                 else
                     output = Path.ChangeExtension(output, ".mp4");
+
+                //读取mp4信息
                 if (result != null && result.Success) 
                 {
                     var data = File.ReadAllBytes(result.ActualFilePath);
@@ -89,6 +106,17 @@ namespace N_m3u8DL_RE.DownloadManager
                     if (info.Scheme != null) Logger.WarnMarkUp($"[grey]Type: {info.Scheme}[/]");
                     if (info.PSSH != null) Logger.WarnMarkUp($"[grey]PSSH(WV): {info.PSSH}[/]");
                     if (info.KID != null) Logger.WarnMarkUp($"[grey]KID: {info.KID}[/]");
+                    //实时解密
+                    if (DownloaderConfig.MP4RealTimeDecryption && streamSpec.Playlist.MediaInit.EncryptInfo.Method != Common.Enum.EncryptMethod.NONE)
+                    {
+                        var enc = result.ActualFilePath;
+                        var dec = Path.Combine(Path.GetDirectoryName(enc)!, Path.GetFileNameWithoutExtension(enc) + "_dec" + Path.GetExtension(enc));
+                        var dResult = await MP4DecryptUtil.DecryptAsync(mp4decrypt, DownloaderConfig.Keys, enc, dec);
+                        if (dResult)
+                        {
+                            result.ActualFilePath = dec;
+                        }
+                    }
                 }
             }
 
@@ -105,7 +133,24 @@ namespace N_m3u8DL_RE.DownloadManager
                 var result = await Downloader.DownloadSegmentAsync(seg, path, headers);
                 FileDic[seg] = result;
                 task.Increment(1);
+                //实时解密
+                if (DownloaderConfig.MP4RealTimeDecryption && seg.EncryptInfo.Method != Common.Enum.EncryptMethod.NONE && result != null) 
+                {
+                    var enc = result.ActualFilePath;
+                    var dec = Path.Combine(Path.GetDirectoryName(enc)!, Path.GetFileNameWithoutExtension(enc) + "_dec" + Path.GetExtension(enc));
+                    var dResult = await MP4DecryptUtil.DecryptAsync(mp4decrypt, DownloaderConfig.Keys, enc, dec, mp4InitFile);
+                    if (dResult)
+                    {
+                        File.Delete(enc);
+                        result.ActualFilePath = dec;
+                    }
+                }
             });
+
+            if (DownloaderConfig.MP4RealTimeDecryption && mp4InitFile != "")
+            {
+                File.Delete(mp4InitFile);
+            }
 
             //校验分片数量
             if (DownloaderConfig.CheckSegmentsCount && FileDic.Values.Any(s => s == null))
@@ -313,30 +358,16 @@ namespace N_m3u8DL_RE.DownloadManager
                 }
             }
 
-            if (DownloaderConfig.Keys != null && DownloaderConfig.Keys.Length > 0) 
+            //调用mp4decrypt解密
+            if (!DownloaderConfig.MP4RealTimeDecryption && DownloaderConfig.Keys != null && DownloaderConfig.Keys.Length > 0) 
             {
-                var APP_DIR = Path.GetDirectoryName(Environment.ProcessPath)!;
-                var fileName = "mp4decrypt";
-                if (Environment.OSVersion.Platform == PlatformID.Win32NT)
-                    fileName += ".exe";
-                var mp4decrypt = Path.Combine(APP_DIR, fileName);
-                if (!File.Exists(mp4decrypt)) mp4decrypt = fileName;
-                if (streamSpec.Playlist!.MediaParts.First().MediaSegments.First().EncryptInfo != null
-                    && streamSpec.Playlist!.MediaParts.First().MediaSegments.First().EncryptInfo.Method != Common.Enum.EncryptMethod.NONE)
+                if (totalCount > 1 && streamSpec.Playlist!.MediaParts.First().MediaSegments.First().EncryptInfo.Method != Common.Enum.EncryptMethod.NONE) 
                 {
                     var enc = output;
                     var dec = Path.Combine(Path.GetDirectoryName(enc)!, Path.GetFileNameWithoutExtension(enc) + "_dec" + Path.GetExtension(enc));
-                    var cmd = string.Join(" ", DownloaderConfig.Keys.Select(k => $"--key {k}")) + $" \"{enc}\" \"{dec}\"";
                     Logger.InfoMarkUp($"[grey]Decrypting...[/]");
-                    await Process.Start(new ProcessStartInfo()
-                    {
-                        FileName = mp4decrypt,
-                        Arguments = cmd,
-                        RedirectStandardOutput = true,
-                        RedirectStandardError = true,
-                        UseShellExecute = false
-                    })!.WaitForExitAsync();
-                    if (File.Exists(dec) && new FileInfo(dec).Length > 0) 
+                    var result = await MP4DecryptUtil.DecryptAsync(mp4decrypt, DownloaderConfig.Keys, enc, dec);
+                    if (result) 
                     {
                         File.Delete(enc);
                         output = dec;
